feat: save and load dates
